refactor(fe): [Issue #98-8-1] features/ ディレクトリへ段階移行
#99 成果物（settlement / home / app）を frontend/src/features/ 配下へ移行。
barrel export を追加し、Screen と App.tsx の import を更新（挙動変更なし）。
