+++ frontend/src/features/app/components/AppViewRouter.tsx
@@ -1,26 +1,26 @@
 import React from 'react';
 import { StyleSheet, View, Text, TouchableOpacity, Platform } from 'react-native';
 import { SafeAreaView } from 'react-native-safe-area-context';
-import { HomeScreen } from '../screens/HomeScreen';
-import HistoryScreen from '../screens/HistoryScreen';
-import { StatisticsScreen } from '../screens/StatisticsScreen';
-import { CategoryManagementScreen } from '../screens/CategoryManagementScreen';
-import { ProductMasterScreen } from '../screens/ProductMasterScreen';
-import { ReceiptScanScreen } from '../screens/ReceiptScanScreen';
-import { PromptEditorScreen } from '../screens/PromptEditorScreen';
-import { AdminStatsScreen } from '../screens/AdminStatsScreen';
-import { AdminMenuScreen } from '../screens/AdminMenuScreen';
-import { SplitEditorScreen } from '../screens/SplitEditorScreen';
-import { SettlementSummaryScreen } from '../screens/SettlementSummaryScreen';
-import { ReceiptTrayScreen } from '../screens/ReceiptTrayScreen';
-import { TotpSettingsScreen } from '../screens/TotpSettingsScreen';
-import { DisplayModeSettings } from './DisplayModeSettings';
-import { DevEnvironmentBanner } from './DevEnvironmentBanner';
-import { theme } from '../theme';
-import { devUiColors, isDevAppEnv } from '../config/appEnv';
+import { HomeScreen } from '../../../screens/HomeScreen';
+import HistoryScreen from '../../../screens/HistoryScreen';
+import { StatisticsScreen } from '../../../screens/StatisticsScreen';
+import { CategoryManagementScreen } from '../../../screens/CategoryManagementScreen';
+import { ProductMasterScreen } from '../../../screens/ProductMasterScreen';
+import { ReceiptScanScreen } from '../../../screens/ReceiptScanScreen';
+import { PromptEditorScreen } from '../../../screens/PromptEditorScreen';
+import { AdminStatsScreen } from '../../../screens/AdminStatsScreen';
+import { AdminMenuScreen } from '../../../screens/AdminMenuScreen';
+import { SplitEditorScreen } from '../../../screens/SplitEditorScreen';
+import { SettlementSummaryScreen } from '../../../screens/SettlementSummaryScreen';
+import { ReceiptTrayScreen } from '../../../screens/ReceiptTrayScreen';
+import { TotpSettingsScreen } from '../../../screens/TotpSettingsScreen';
+import { DisplayModeSettings } from '../../../components/DisplayModeSettings';
+import { DevEnvironmentBanner } from '../../../components/DevEnvironmentBanner';
+import { theme } from '../../../theme';
+import { devUiColors, isDevAppEnv } from '../../../config/appEnv';
 import type { AppViewType } from '../hooks/useAppSession';
-import type { ReceiptScanInitialData } from '../types/receiptScan';
-import type { ReceiptForSplitEditor } from '../types/settlement';
+import type { ReceiptScanInitialData } from '../../../types/receiptScan';
+import type { ReceiptForSplitEditor } from '../../../types/settlement';
 
 export interface AppViewRouterProps {
   currentView: AppViewType;
