+++ frontend/src/features/home/index.ts
@@ -1,0 +1,2 @@
+export { useHomeDashboard } from './hooks/useHomeDashboard';
+export { useReceiptUpload } from './hooks/useReceiptUpload';
