+++ frontend/src/features/settlement/components/SettlementDetailTable.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { StyleSheet, Text, View } from 'react-native';
-import { theme, tableStyles } from '../../theme';
-import type { SettlementMemberSummary } from '../../types/settlement';
+import { theme, tableStyles } from '../../../theme';
+import type { SettlementMemberSummary } from '../../../types/settlement';
 
 interface SettlementDetailTableProps {
   summaryData: SettlementMemberSummary[];
