+++ frontend/src/features/settlement/components/SettlementMemberCards.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { StyleSheet, Text, View } from 'react-native';
-import { theme } from '../../theme';
-import type { SettlementMemberSummary } from '../../types/settlement';
+import { theme } from '../../../theme';
+import type { SettlementMemberSummary } from '../../../types/settlement';
 
 interface SettlementMemberCardsProps {
   summaryData: SettlementMemberSummary[];
